JavaScript-DOM
When you click on the color palette icon located at the top right of the page, a left side ribbon appears. #Bootstrap5.2

When changing color, a JavaScript code must directly modify the DOM of the "index.html" page #querySelector #eventListener #change #innerHTML
A "Default Colors" button resets the page to its original colors.
The page has an arrow icon in its footer. A click on this icon will bring the scroll back to the top of the page. #scrollTo #smooth
In addition, cookies are used to store color values.